Add license and cleanup main.swift
